Quiz _game
First, we need to decide on the type of quiz game we want to create. we can choose from multiple-choice questions.
Next, we need to create a list of questions and answers for your quiz game.
Then, we can use Python’s built-in functions like input() and print() to interact with the user and display the questions and answers.
we can also use Python’s random module to randomize the order of the questions.
Finally, we can keep track of the user’s score and display it at the end of the quiz.